+++ packages/keyring-sdk/src/keyring-account-registry.ts
@@ -1,7 +1,7 @@
 import type { AccountId } from '@metamask/keyring-utils';
 import { v4 as uuidv4 } from 'uuid';
 
-import type { KeyringAccount } from '../../account';
+import type { KeyringAccount } from '@metamask/keyring-api';
 
 /**
  * In-memory registry for KeyringAccount objects.